[Feat] #105 AuthService
