+++ src/main/java/com/umust/dobonglife/domain/auth/controller/AuthController.java
@@ -4,6 +4,7 @@ import com.umust.dobonglife.domain.auth.controller.dto.request.GoogleLoginReques
 import com.umust.dobonglife.domain.auth.controller.dto.request.KakaoLoginRequest;
 import com.umust.dobonglife.domain.auth.controller.dto.request.RefreshTokenRequest;
 import com.umust.dobonglife.domain.auth.controller.dto.response.TokenResponse;
+import com.umust.dobonglife.domain.auth.service.AuthService;
 import com.umust.dobonglife.domain.auth.service.JwtService;
 import com.umust.dobonglife.domain.auth.service.KakaoAuthService;
 import com.umust.dobonglife.global.auth.resolver.CurrentUserId;
@@ -35,6 +36,7 @@ public class AuthController {
     private final JwtService jwtService;
     private final GoogleAuthService googleAuthService;
     private final KakaoAuthService kakaoAuthService;
+    private final AuthService authService;
 
     @Operation(summary = "카카오 로그인", description = "카카오 로그인을 합니다.")
     @ApiResponse(
@@ -63,7 +65,7 @@ public class AuthController {
     )
     @PostMapping("/logout")
     public BaseResponse<Void> logout(HttpServletRequest request){
-        jwtService.logout(request);
+        authService.logout(request);
         return BaseResponse.ok(null);
     }
